refactor: extract PredictionMarketsPage data logic into usePredictionMarketsData hook (F5)
Move the market-list/portfolio/analysis fetch and state into a page-local
hook (components/prediction-markets/usePredictionMarketsData.js).
PredictionMarketsPage becomes a thin render shell (375 -> 237 lines).

Behavior-preserving: hook body moved verbatim (same hook order incl.
useAuth, same effect deps), no early returns, returning exactly the state
and handlers the JSX consumes. Verified by the existing 5-test suite, full
suite (83), lint, and build.
